backend/assistant_service/requirements.txt
faiss-cpu
sentence-transformers
numpy
SpeechRecognition
pyttsx3
pyaudio
deep-translator
gTTS
pygame
flask
flask-cors
requests
